Create Lab 2-unsorted
In this project, you will implement a set abstract data type for strings. Your interface and implementation must bekept separate. Separate source files that providemainwill be provided for testing your data type.

ou will write two different implementations of the data type for this assignment.  First, implement a set using anunsorted array of lengthm>0, in which the firstn≤mslots are used to holdnstrings in some arbitrary order. Usesequential search to locate an element in the array.  Second, implement a set using a sorted array of lengthm>0,in which the firstn≤mslots are used to holdnstrings in ascending order. Use binary search to locate an elementin the array.For both implementations, rather than duplicating the search logic in several functions, write an auxiliary func-tionsearchthat returns the location of an element in an array. Declaresearchas static so it is not visible outsidethe source file.   Usesearchto implement the functions in the interface.   Your implementation should allocatememory and copy the string when adding, and therefore also deallocate memory when removing.By the end of the first lab, you should have finished the first implementation.  You should verify that the ADTworks with both test programs (unique, and thenparity) on all of the files.  Note thatuniquetakes an optionalsecond  file,  whose  words  itremovesfrom  the  set,  thus  allowing  you  to  test  insertion  followed  by  deletion.   Incontrast,parityinterweaves  insertion  and  deletion,  so  it  is  a  tougher  test.   By  the  end  of  the  second  lab,  youshould have finished both implementations.  Before the due date, make sure you finish commenting and testingyour code. You should use theassertmacro defined in<assert.h>where appropriate.