Create Lab 5- Pqueue.c
Well, Professor Gosheim Loony has done it again!  He has written part of a project, but did not have time to finishit, so he has given you that task.  Professor Loony has been working on a file compression project using Huffmancoding. This time around, he has written the function to perform the actual compression given the Huffman tree,but still needs to construct the tree. (Are you really surprised that Professor Loony has leftyouup a tree?)The tree data type itself is easy to construct as it consists simply of a parent pointer and a value. Therefore, Pro-fessor Loony does not require you write a tree ADT type. However, since one of the key components of the Huffmancoding algorithm is a priority queue, he does require you to write a priority queue ADT for this assignment.

As required by Professor Loony, you will write a priority queue abstract data type using a binary heap implementedusing an array.  Operations to add and remove entries are required to run inO(logn) time.  Unlike previous datatypes that used arrays, thecreateQueuefunction does not require the maximum number of values as a parameter.Instead, Professor Loony wants you to increase the length of your array when it reaches capacity so it dynamicallyincreases to accommodate the number of values. Thus, we eliminate the need for the client to inform us how largeof an array it requires.Fortunately, this technique is easy in C using the functionrealloc, which reallocates an array to a given sizeand  returns  a  pointer  to  the  new  array.   However,  reallocation  can  be  expensive,  as  it  may  require  allocating  acompletely new array and then copying the data from the old array to the new array.  Therefore, we don’t want todo it too often.  A common technique is to start the array at a small size (we’ll use ten elements).  When the arrayreaches capacity (i.e., the number of stored values equals the length of the array), then double the array’s lengthusingrealloc.  Doubling the array’s length may seem excessive, but it guarantees that for the nextninsertions,we won’t need to do a reallocation.  So, although we may have had to copynvalues in a reallocation, the nextnvalues inserted won’t require a reallocation, so theamortizedcost over time isO(n)/O(n)=O(1).  (Note that thistechnique is not specific to a binary heap, but works with any array, so we could go back and reimplement ourunsorted and sorted array implementations in the same manner.)